Feedback resolved
