+++ public/react/SettingsSidebar/components/IssueSource/components/StatusSelect/ExcludedStatusSelect.tsx
@@ -1,9 +1,20 @@
 import React, { FC, useId } from "react";
 import Select from "@atlaskit/select";
 import { Label } from "@atlaskit/form";
-import { ExcludedStatusSelectProps } from "./types";
 import useExcludedStatusSelect from "./hooks/useExcludedStatusSelect";
+import type { MultiValue } from "react-select";
 
+export interface ExcludedStatusSelectOption {
+  label: string;
+  value: string;
+}
+
+export interface ExcludedStatusSelectProps {
+  label: string;
+  placeholder: string;
+  value: MultiValue<ExcludedStatusSelectOption>;
+  onChange?: (value: MultiValue<ExcludedStatusSelectOption>) => void;
+}
 const ExcludedStatusSelect: FC<ExcludedStatusSelectProps> = ({
   label,
   placeholder,
